finish project

+++ src/app/services/authentication/authentication.service.ts
@@ -8,13 +8,23 @@ export class AuthenticationService {
 
   saveUser(user: any) {
     let users = JSON.parse(localStorage.getItem('users') || '[]');
+  
+    const userExists = users.some((u: any) => u.email === user.email);
+  
+    if(!userExists) {
+      users.push(user);
+      localStorage.setItem('users', JSON.stringify(users));
+      return true
+    }
 
-    users.push(user);
-
-    localStorage.setItem('users', JSON.stringify(users));
+    return false
   }
 
-  getUser() {
-    
+  getUser(email: string, password: string): boolean {
+    let users = JSON.parse(localStorage.getItem('users') || '[]');
+  
+    const user = users.find((u: any) => u.email === email && u.password === password);
+  
+    return !!user;
   }
 }
